requirements.txt
appdirs==1.4.3
apsw==3.8.2.post1
apturl==0.5.2
asgiref==3.2.10
attrs==19.3.0
autobahn==17.10.1
autograd==1.3
Automat==0.8.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1
django-filter==2.4.0
django-heroku==0.3.1
django-rest-authemail==2.0.6
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.2
dnspython==1.16.0
docutils==0.16
duplicity==0.8.12.0
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
feedparser==5.2.1
filelock==3.0.12
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
ifaddr==0.1.6
importlib-metadata==1.5.0
incremental==16.10.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.2
mechanize==0.4.5
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
msgpack==0.6.2
netifaces==0.10.4
nose==1.3.7
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pbr==5.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
platformdirs==2.2.0
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.9.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pychm==0.8.6
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2019.8.19
reportlab==3.5.34
repoze.lru==0.7
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rosdep-modules==0.20.1
rosdistro==0.8.3
rosdistro-modules==0.8.3
rospkg==1.3.0
rospkg-modules==1.3.0
Routes==2.4.1
scipy==1.6.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smmap==4.0.0
soupsieve==1.9.5
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
tinyik==2.2.0
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.7.2
virtualenv-clone==0.3.0
wadllib==1.3.3
webencodings==0.5.1
WebOb==1.8.5
whitenoise==5.3.0
wsaccel==0.6.2
xkit==0.0.0
zeroconf==0.24.4
zipp==1.0.0
zope.interface==4.7.1
